Move pkg loading and sorting logic to gnomod
