+++ gno.land/cmd/gnoland/main.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
@@ -35,12 +34,6 @@ type gnolandCfg struct {
 	chainID               string
 	genesisRemote         string
 	rootDir               string
-}
-
-type pkg struct {
-	name     string
-	path     string
-	requires []string
 }
 
 func main() {
@@ -215,18 +208,17 @@ func makeGenesisDoc(
 	txs := []std.Tx{}
 
 	// List initial packages to load from examples.
-	pkgs, err := listGnoPkgs(filepath.Join("..", "examples"))
+	pkgs, err := gnomod.ListPkgs(filepath.Join("..", "examples"))
 	if err != nil {
 		panic(fmt.Errorf("listing gno packages: %w", err))
 	}
-	if err := sortPkgs(pkgs); err != nil {
+	if err := gnomod.SortPkgs(pkgs); err != nil {
 		panic(fmt.Errorf("sorting packages: %w", err))
 	}
 
 	for _, pkg := range pkgs {
 		// open files in directory as MemPackage.
-		fmt.Println("adding package", pkg.path, pkg.name)
-		memPkg := gno.ReadMemPackage(pkg.path, pkg.name)
+		memPkg := gno.ReadMemPackage(pkg.Path(), pkg.Name())
 		var tx std.Tx
 		tx.Msgs = []std.Msg{
 			vmm.MsgAddPackage{
@@ -308,102 +300,4 @@ func loadGenesisBalances(path string) []string {
 		balances = append(balances, line)
 	}
 	return balances
-}
-
-// listGnoPkgs lists all gno packages in the given root directory.
-func listGnoPkgs(root string) ([]pkg, error) {
-	var pkgs []pkg
-
-	err := filepath.WalkDir(root, func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if d.IsDir() {
-			goModPath := filepath.Join(path, "gno.mod")
-			data, err := os.ReadFile(goModPath)
-			if err == nil {
-				gnoMod, err := gnomod.Parse(goModPath, data)
-				if err != nil {
-					return fmt.Errorf("parse: %w", err)
-				}
-				gnoMod.Sanitize()
-				if err := gnoMod.Validate(); err != nil {
-					return fmt.Errorf("validate: %w", err)
-				}
-				pkgs = append(pkgs, pkg{
-					name: gnoMod.Module.Mod.Path,
-					path: path,
-					requires: func() []string {
-						var reqs []string
-						for _, req := range gnoMod.Require {
-							reqs = append(reqs, req.Mod.Path)
-						}
-						return reqs
-					}(),
-				})
-				return fs.SkipDir
-			} else if !os.IsNotExist(err) {
-				return err
-			}
-		}
-
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return pkgs, nil
-}
-
-// sortPkgs sorts the given packages by their dependencies.
-func sortPkgs(pkgs []pkg) error {
-	visited := make(map[string]bool)
-	onStack := make(map[string]bool)
-	sortedPkgs := make([]pkg, 0, len(pkgs))
-
-	var visit func(pkg pkg) error
-	visit = func(pkg pkg) error {
-		if onStack[pkg.name] {
-			return fmt.Errorf("cycle detected: %s", pkg.name)
-		}
-		if visited[pkg.name] {
-			return nil
-		}
-
-		visited[pkg.name] = true
-		onStack[pkg.name] = true
-
-		// Visit package's dependencies
-		for _, req := range pkg.requires {
-			found := false
-			for _, p := range pkgs {
-				if p.name == req {
-					if err := visit(p); err != nil {
-						return err
-					}
-					found = true
-					break
-				}
-			}
-			if !found {
-				return fmt.Errorf("missing dependency '%s' for package '%s'", req, pkg.name)
-			}
-		}
-
-		onStack[pkg.name] = false
-		sortedPkgs = append(sortedPkgs, pkg)
-		return nil
-	}
-
-	// Visit all packages
-	for _, p := range pkgs {
-		if err := visit(p); err != nil {
-			return err
-		}
-	}
-
-	copy(pkgs, sortedPkgs)
-	return nil
 }
